api-service/requirements.txt
fastapi
slowapi
uvicorn
psycopg2-binary
python-dotenv
requests